refactro(server): job dto
