--- server/src/domain/job/dto/index.ts
@@ -1,2 +1,0 @@
-export * from './job-command.dto';
-export * from './job-id.dto';
