--- server/src/domain/job/dto/job-id.dto.ts
@@ -1,10 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsEnum, IsNotEmpty } from 'class-validator';
-import { QueueName } from '../job.constants';
-
-export class JobIdDto {
-  @IsNotEmpty()
-  @IsEnum(QueueName)
-  @ApiProperty({ type: String, enum: QueueName, enumName: 'JobName' })
-  id!: QueueName;
-}
