+++ server/src/domain/job/index.ts
@@ -1,6 +1,5 @@
-export * from './dto';
 export * from './job.constants';
+export * from './job.dto';
 export * from './job.interface';
 export * from './job.repository';
 export * from './job.service';
-export * from './response-dto';
