--- server/src/domain/job/response-dto/index.ts
@@ -1,1 +1,0 @@
-export * from './all-job-status-response.dto';
